integrate ml solid into pipeline and fix tests
Summary:
This revision integrates the train_lstm_model solid into the pipeline and fix tests.

Call out: The whole upload a training set to the cloud solid doesn't feel right. It seems we are just rebuilding intermediates. The original use case was that we wanted to have 2 pipelines but decided to merge it to one. Thoughts?

Test Plan: unit

Differential Revision: https://dagster.phacility.com/D2082