Include exception message in logs
Summary:
We previously added exception error messages as a log key `message`, but we skip the "message" key when printing logs since it has duplicated information for other event types.

This made it hard to debug exceptions in solids since users cannot see the actual exception they are hitting.

This diff renames the log key to "error_message" and prints the exception message inline with the stack.

Test Plan: unit

Reviewers: #ft

Differential Revision: https://dagster.phacility.com/D2088